+++ R/data_sad_nca.R
@@ -1,5 +1,5 @@
 #' Example NCA Parameter Dataset Output from PKNCA
-#' @format ## `data.frame` a data frame with 720 rows and 23 columns:
+#' @format ## `data.frame` a data frame with 648 rows and 11 columns:
 #' \describe{
 #'   \item{ID}{NONMEM subject identifier}
 #'   \item{DOSE}{Nominal dose assignment (units: mg)}
@@ -12,7 +12,6 @@
 #'   \item{units_dose}{Dose units}
 #'   \item{units_conc}{Concentration units}
 #'   \item{units_time}{Time units}
-#'   ...
 #' }
 #' @source simulated
 "data_sad_nca"