+++ data-raw/logo_raw.R
@@ -27,7 +27,7 @@ vpc_plot <- plot_vpc_cont(
   sim = simout,
   pcvpc = TRUE,
   pi = c(0.05, 0.95),
-  ci = c(0.05, 0.95),
+  ci = 0.90,
   log_y = TRUE,
   vpc_theme = vpc_theme)+
   theme_classic()+